test(popver): migrate nested
