+++ core/src/components/popover/test/nested/popover.e2e.ts-snapshots/popover-nested-md-ltr-Mobile-Safari-linux.png
